refactor:커스텀 예외 클래스 추가

+++ src/main/java/com/study/sjicnho/delivery/user/service/UserService.java
@@ -1,13 +1,15 @@
 package com.study.sjicnho.delivery.user.service;
 
-import com.study.sjicnho.delivery.user.RoleType;
+import com.study.sjicnho.delivery.ErrorCode;
+import com.study.sjicnho.delivery.store.exception.NoSuchStore;
+import com.study.sjicnho.delivery.user.entity.UserRole;
+import com.study.sjicnho.delivery.user.exception.NoSuchUser;
 import com.study.sjicnho.delivery.user.repository.UserRepository;
 import com.study.sjicnho.delivery.user.dto.UserDto;
 import com.study.sjicnho.delivery.user.entity.User;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
-
 import java.util.List;
 
 @Service
@@ -22,42 +24,27 @@ public class UserService{
         this.bCryptPasswordEncoder = bCryptPasswordEncoder;
     }
 
-    public void signUpProcess(UserDto userDto) {
+    public User signUp(UserDto userDto) {
 
         String username = userDto.getEmail();
         String password = userDto.getPassword();
-        RoleType role = userDto.getRoleType();
+        String name = userDto.getName();
+        UserRole role = userDto.getUserRole();
 
         Boolean isExist = userRepository.existsByEmail(username);
 
         if (isExist) {
-            throw new IllegalStateException("이미 가입되어 있는 이메일입니다.");
+            throw new RuntimeException("이미 가입되어 있는 이메일입니다.");
         }
 
         userDto.setEmail(username);
         userDto.setPassword(bCryptPasswordEncoder.encode(password));
-        userDto.setRoleType(role);
+        userDto.setUserRole(role);
+        userDto.setName(name);
         User data = userDto.toEntity();
 
-        userRepository.save(data);
+        return userRepository.save(data);
     }
-
-//    @Override
-//    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-//        Boolean isExist = userRepository.existsByUserEmail(email);
-//
-//        if (!isExist) {
-//            throw new UsernameNotFoundException(email);
-//        }
-//        log.info("{user} :", isExist);
-//
-//        return org.springframework.security.core.userdetails.User.builder()
-//                .username(user.getEmail())
-//                .password(user.getPassword())
-//                .roles(user.getRoleType().toString())
-//                .build();
-//
-//    }
 
 
     public List<User> getAll() {
@@ -65,33 +52,9 @@ public class UserService{
     }
 
     public User getById(Integer id) {
-        return userRepository.findById(id).orElse(null);
+        return userRepository.findById(id)
+                .orElseThrow(() -> new NoSuchUser(ErrorCode.USER_NOT_FOUND));
     }
-
-//    public User getByEmail(String email){
-//        return userRepository.findByUserEmail(email);
-//    }
-//
-//    public User signUp(UserDto userDto) {
-//
-//        //비밀번호 암호화
-//        userDto.setPassword(passwordEncoder.encode(userDto.getPassword()));
-//
-//        //DTO->Entity
-//        User user = userDto.toEntity();
-//
-//        validDuplicateUser(user);
-//
-//        return userRepository.save(user);
-//    }
-
-//    //중복 회원 유효성 검사
-//    private void validDuplicateUser(User user){
-//        User findUser = userRepository.findByUserEmail(user.getEmail());
-//        if(findUser != null){
-//            throw new IllegalStateException("이미 가입된 회원입니다.");
-//        }
-//    }
 
     public void update(Integer id, UserDto dto) {
 
@@ -100,30 +63,20 @@ public class UserService{
         //DTO->Entity
         User user = dto.toEntity();
 
-        User target = userRepository.findById(id).orElseThrow(null);
+        User target = userRepository.findById(id)
+                .orElseThrow(() -> new NoSuchUser(ErrorCode.USER_NOT_FOUND));
+
         if(target != null){
             userRepository.save(user);
         }
     }
 
     public void delete(Integer id) {
-        userRepository.deleteById(id);
+
+        User target = userRepository.findById(id)
+                .orElseThrow(() -> new NoSuchUser(ErrorCode.USER_NOT_FOUND));
+        if(target != null){
+            userRepository.deleteById(id);
+        }
     }
-
-//    public User login(UserDto userDto) {
-//
-//        String email = userDto.getEmail();
-//        log.info("{email}:", email);
-//
-//        User user = userRepository.findByEmail(userDto.getEmail());
-//        log.info("{user}:", user);
-//
-//        if (!userDto.getEmail().equals(user.getEmail())) {
-//            throw new IllegalStateException("이메일을 확인해주세요");
-//        } else {
-//            return user;
-//        }
-//
-//    }
-
 }
